refactor(vision): extract AudioEncoder collaborator out of vision_module (slice 8)
Audio doesn't belong in a 'vision' module. Moves the audio machinery verbatim into a
dedicated AudioEncoder (audio_encoder.hpp/.cpp): AudioEncoderRuntime, init_audio_profile,
init_whisper_audio_profile, select_audio_profile, encode_block_fp32,
try_pick_tokens_by_output_bytes + Gemma4/Whisper Init + Encode. VisionModule now holds a
unique_ptr<AudioEncoder>, selected via adapter_->audioKind() at Init and called in the
AUDIO branch; the multimodal injection (audio_pad placeholders) stays in VisionModule.
Also factors the backend device-copy infra (ax_runner_t + v_h2d/v_d2h + V_WADDR/V_RADDR)
into vision_runtime.hpp, shared by vision_module.cpp + audio_encoder.cpp.

vision_module.cpp type_==VLMType: audio dispatch -> 0 (2 residual: None-disabled check +
classic input_is_nchw validation guard, both non-dispatch).

Behavior-preserving move. Verified AX650N .51, build_650 + build_x86 both compile+link;
golden --compare BYTE-IDENTICAL for 5 IMAGE + 3 VIDEO (non-audio unaffected). Audio path
is a verbatim relocation, not runtime-verified (no Gemma4/Qwen3Omni board on hand).
